Moved project list to its own component category since it may be reused for other pages.
